requirements.txt
Pillow>=10.0.0
Flask>=3.0.0
waitress>=2.1.2
